more tests minor refactor
